merging from the main fork.

+++ tests/TestCase/Shell/I18nShellTest.php
@@ -1,21 +1,19 @@
 <?php
 /**
- * CakePHP(tm) : Rapid Development Framework (http://cakephp.org)
- * Copyright (c) Cake Software Foundation, Inc. (http://cakefoundation.org)
+ * CakePHP(tm) : Rapid Development Framework (https://cakephp.org)
+ * Copyright (c) Cake Software Foundation, Inc. (https://cakefoundation.org)
  *
  * Licensed under The MIT License
  * For full copyright and license information, please see the LICENSE.txt
  * Redistributions of files must retain the above copyright notice.
  *
- * @copyright     Copyright (c) Cake Software Foundation, Inc. (http://cakefoundation.org)
- * @link          http://cakephp.org CakePHP(tm) Project
+ * @copyright     Copyright (c) Cake Software Foundation, Inc. (https://cakefoundation.org)
+ * @link          https://cakephp.org CakePHP(tm) Project
  * @since         3.0.8
- * @license       http://www.opensource.org/licenses/mit-license.php MIT License
+ * @license       https://opensource.org/licenses/mit-license.php MIT License
  */
 namespace Cake\Test\TestCase\Shell;
 
-use Cake\Cache\Cache;
-use Cake\Datasource\ConnectionManager;
 use Cake\Shell\I18nShell;
 use Cake\TestSuite\TestCase;
 
@@ -33,7 +31,7 @@ class I18nShellTest extends TestCase
     public function setUp()
     {
         parent::setUp();
-        $this->io = $this->getMock('Cake\Console\ConsoleIo');
+        $this->io = $this->getMockBuilder('Cake\Console\ConsoleIo')->getMock();
         $this->shell = new I18nShell($this->io);
 
         $this->localeDir = TMP . 'Locale' . DS;
@@ -48,13 +46,16 @@ class I18nShellTest extends TestCase
     {
         parent::tearDown();
 
-        $deDir = $this->localeDir . 'de' . DS;
+        $deDir = $this->localeDir . 'de_DE' . DS;
 
-        unlink($this->localeDir . 'default.pot');
-        unlink($this->localeDir . 'cake.pot');
-
-        unlink($deDir . 'default.po');
-        unlink($deDir . 'cake.po');
+        if (file_exists($this->localeDir . 'default.pot')) {
+            unlink($this->localeDir . 'default.pot');
+            unlink($this->localeDir . 'cake.pot');
+        }
+        if (file_exists($deDir . 'default.po')) {
+            unlink($deDir . 'default.po');
+            unlink($deDir . 'cake.po');
+        }
     }
 
     /**
@@ -64,7 +65,7 @@ class I18nShellTest extends TestCase
      */
     public function testInit()
     {
-        $deDir = $this->localeDir . 'de' . DS;
+        $deDir = $this->localeDir . 'de_DE' . DS;
         if (!is_dir($deDir)) {
             mkdir($deDir, 0770, true);
         }
@@ -77,10 +78,10 @@ class I18nShellTest extends TestCase
             unlink($deDir . 'cake.po');
         }
 
-        $this->shell->io()->expects($this->at(0))
+        $this->shell->getIo()->expects($this->at(0))
             ->method('ask')
-            ->will($this->returnValue('de'));
-        $this->shell->io()->expects($this->at(1))
+            ->will($this->returnValue('de_DE'));
+        $this->shell->getIo()->expects($this->at(1))
             ->method('ask')
             ->will($this->returnValue($this->localeDir));
 
@@ -89,5 +90,18 @@ class I18nShellTest extends TestCase
 
         $this->assertFileExists($deDir . 'default.po');
         $this->assertFileExists($deDir . 'cake.po');
+    }
+
+    /**
+     * Test that the option parser is shaped right.
+     *
+     * @return void
+     */
+    public function testGetOptionParser()
+    {
+        $this->shell->loadTasks();
+        $parser = $this->shell->getOptionParser();
+        $this->assertArrayHasKey('init', $parser->subcommands());
+        $this->assertArrayHasKey('extract', $parser->subcommands());
     }
 }
